feat(passkeys): move form submission logic to formUtils.js
