+++ frontend/passkeys.js
@@ -1,5 +1,6 @@
 import DOMPurify from 'dompurify';
 import M from 'materialize-css';
+import { createAndSubmitForm } from './formUtils.js';
 import { getPasskeyNameFromUser } from './passkeyNameModal.js';
 
 /**
@@ -211,28 +212,6 @@ export function setUpDeletePasskeyButtons(selector) {
             }
         });
     });
-}
-
-/**
- * Creates and submits a form with the provided data
- * @param {string} targetHref - Form submission URL
- * @param {Object} formData - Data to include in the form
- */
-function createAndSubmitForm(targetHref, formData) {
-    const form = document.createElement('form');
-    form.setAttribute('method', 'post');
-    form.setAttribute('action', DOMPurify.sanitize(targetHref));
-
-    Object.entries(formData).forEach(([name, value]) => {
-        const input = document.createElement('input');
-        input.setAttribute('type', 'hidden');
-        input.setAttribute('name', DOMPurify.sanitize(name));
-        input.setAttribute('value', DOMPurify.sanitize(value));
-        form.appendChild(input);
-    });
-
-    document.body.append(form);
-    form.submit();
 }
 
 /**
